Comunicazione con i client. Il server è in grado di avviare la partita inviando le posizioni iniziali dei giocatori e ricevere dati dai client.
PS: In sviluppo il passaggio del JSON contenente le coordinate di un client all'altro.

+++ src/main/java/org/example/Player.java
@@ -2,13 +2,13 @@ package org.example;
 
 public class Player {
     private int nHeart = 3;
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
     public Player(int nHeart, double x, double y) {
         this.nHeart = nHeart;
-        this.x = x;
-        this.y = y;
+        this.x = (int) x;
+        this.y = (int) y;
     }
 
     public int getnHeart() {
@@ -24,7 +24,7 @@ public class Player {
     }
 
     public void setX(double x) {
-        this.x = x;
+        this.x = (int) x;
     }
 
     public double getY() {
@@ -32,6 +32,6 @@ public class Player {
     }
 
     public void setY(double y) {
-        this.y = y;
+        this.y = (int) y;
     }
 }
